Drop sequences, update flu samples and v/m ingestion

+++ DB/structure/sql/samples_lineages/create_table_samples_lineages.sql
@@ -1,6 +1,6 @@
 create table samples_lineages (
 	id bigserial not null,
-	sample_id bigint not null,
+	sample_id integer not null,
 	lineage_id bigint not null,
 	abundance double precision,
 	is_consensus_call boolean not null
